Fix adding percentage 50% Adding exp bar in minecraft and level

+++ src/main/java/com/thomas7520/saofrance/leveling/player/IPlayerLeveling.java
@@ -15,9 +15,9 @@ public interface IPlayerLeveling {
 
     int getExperience();
 
-    int getPercentageExperience();
+    double getPercentageExperience();
 
-    int getPercentageExperienceRequire();
+    double getPercentageExperienceRequire();
 
     boolean canLevelUp(int experienceAcquire);
